examples/requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
ordereddict==1.1
python-dotenv==0.5.1
Werkzeug==0.11.10
